Move package from tests/workspace/common to pkg/integration/common
